Break metal into smaller chunks
